+++ myems-admin/app/controllers/settings/workingcalendar/workingcalendar.master.controller.js
@@ -1,0 +1,10 @@
+'use strict';
+
+app.controller('WorkingCalendarMasterController', function($scope) {
+
+
+	$scope.$on('handleEmitWorkingCalendarChanged', function(event) {
+		$scope.$broadcast('handleBroadcastWorkingCalendarChanged');
+	});
+
+});